% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clasificacion_dataf.R
\name{clasificacion}
\alias{clasificacion}
\title{Clasificacion columnas dataframe}
\usage{
clasificacion(dataframe1)
}
\arguments{
\item{dataframe1}{Dataframe exportado ó creado con datos externos}
}
\value{
clase de cada columna
}
\description{
Clasificacion columnas dataframe
}
\examples{
d <- data.frame(Nombre = c("Esteban", "Mario", "Beatriz"),
                Edad = c(2, 93, 57),
                Ciudad = factor(c("Jaén", "Granada", "Jaén")),
                Casado = c(FALSE, TRUE, FALSE))
clasificacion(d)

edad <- c(22, 34, 29, 25, 30)
tiempo <- c(14.21, 10.36, 11.89, 13.81, 12.03)
sexo <- c("M","H","H","M","M")
misDatos <- data.frame(edad,tiempo,sexo)
clasificacion(misDatos)

}
